+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Atoms can be now used as attribute keys
 
 ## [0.1.1] - 2017-07-26
 ### Changed
@@ -16,4 +18,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased]: https://github.com/SoftwareMansion/elixir-xattr/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/SoftwareMansion/elixir-xattr/compare/v0.1.0...v0.1.1
-